+++ ExternalHelpers/WindowsMasterVolume/AudioManager.cs
@@ -1,0 +1,321 @@
+﻿using System.Runtime.InteropServices;
+
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    /// <summary>
+    /// Controls audio using the Windows CoreAudio API
+    /// from: http://stackoverflow.com/questions/14306048/controling-volume-mixer
+    /// </summary>
+    public static class AudioManager
+    {
+        public static int GetMasterVolume(double roundValue)
+        {
+            return (int)(Math.Round(GetMasterVolume() / roundValue) * roundValue);
+        }
+
+        #region Master Volume Manipulation
+
+        /// <summary>
+        /// Gets the current master volume in scalar values (percentage)
+        /// </summary>
+        /// <returns>-1 in case of an error, if successful the value will be between 0 and 100</returns>
+        public static int GetMasterVolume()
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol != null)
+                {
+                    float volumeLevel;
+                    masterVol.GetMasterVolumeLevelScalar(out volumeLevel);
+                    return (int)(volumeLevel * 100);
+                }
+
+                return -1;
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        /// <summary>
+        /// Gets the mute state of the master volume. 
+        /// While the volume can be muted the <see cref="GetMasterVolume"/> will still return the pre-muted volume value.
+        /// </summary>
+        /// <returns>false if not muted, true if volume is muted</returns>
+        public static bool GetMasterVolumeMute()
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol == null)
+                    return false;
+
+                bool isMuted;
+                masterVol.GetMute(out isMuted);
+                return isMuted;
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        /// <summary>
+        /// Sets the master volume to a specific level
+        /// </summary>
+        /// <param name="newLevel">Value between 0 and 100 indicating the desired scalar value of the volume</param>
+        public static void SetMasterVolume(int newLevel)
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol == null)
+                    return;
+
+                masterVol.SetMasterVolumeLevelScalar((float)newLevel / 100.0f, Guid.Empty);
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        /// <summary>
+        /// Increments or decrements the current volume level by the <see cref="stepAmount"/>.
+        /// </summary>
+        /// <param name="stepAmount">Value between -100 and 100 indicating the desired step amount. Use negative numbers to decrease
+        /// the volume and positive numbers to increase it.</param>
+        /// <returns>the new volume level assigned</returns>
+        public static float StepMasterVolume(float stepAmount)
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol == null)
+                    return -1;
+
+                float stepAmountScaled = stepAmount / 100;
+
+                // Get the level
+                float volumeLevel;
+                masterVol.GetMasterVolumeLevelScalar(out volumeLevel);
+
+                // Calculate the new level
+                float newLevel = volumeLevel + stepAmountScaled;
+                newLevel = Math.Min(1, newLevel);
+                newLevel = Math.Max(0, newLevel);
+
+                masterVol.SetMasterVolumeLevelScalar(newLevel, Guid.Empty);
+
+                // Return the new volume level that was set
+                return newLevel * 100;
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        /// <summary>
+        /// Mute or unmute the master volume
+        /// </summary>
+        /// <param name="isMuted">true to mute the master volume, false to unmute</param>
+        public static void SetMasterVolumeMute(bool isMuted)
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol == null)
+                    return;
+
+                masterVol.SetMute(isMuted, Guid.Empty);
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        /// <summary>
+        /// Switches between the master volume mute states depending on the current state
+        /// </summary>
+        /// <returns>the current mute state, true if the volume was muted, false if unmuted</returns>
+        public static bool ToggleMasterVolumeMute()
+        {
+            IAudioEndpointVolume? masterVol = null;
+            try
+            {
+                masterVol = GetMasterVolumeObject();
+                if (masterVol == null)
+                    return false;
+
+                bool isMuted;
+                masterVol.GetMute(out isMuted);
+                masterVol.SetMute(!isMuted, Guid.Empty);
+
+                return !isMuted;
+            }
+            finally
+            {
+                if (masterVol != null)
+                    Marshal.ReleaseComObject(masterVol);
+            }
+        }
+
+        private static IAudioEndpointVolume? GetMasterVolumeObject()
+        {
+            IMMDeviceEnumerator? deviceEnumerator = null;
+            IMMDevice? speakers = null;
+            try
+            {
+                deviceEnumerator = (IMMDeviceEnumerator)(new MMDeviceEnumerator());
+                deviceEnumerator.GetDefaultAudioEndpoint(EDataFlow.eRender, ERole.eMultimedia, out speakers);
+
+                Guid IID_IAudioEndpointVolume = typeof(IAudioEndpointVolume).GUID;
+                object? o = null;
+                speakers?.Activate(ref IID_IAudioEndpointVolume, 0, IntPtr.Zero, out o);
+                IAudioEndpointVolume? masterVol = o == null ? null : (IAudioEndpointVolume)o;
+
+                return masterVol;
+            }
+            finally
+            {
+                if (speakers != null) Marshal.ReleaseComObject(speakers);
+                if (deviceEnumerator != null) Marshal.ReleaseComObject(deviceEnumerator);
+            }
+        }
+
+        #endregion
+
+        #region Individual Application Volume Manipulation
+
+        public static float? GetApplicationVolume(int pid)
+        {
+            ISimpleAudioVolume? volume = GetVolumeObject(pid);
+            if (volume == null)
+                return null;
+
+            float level;
+            volume.GetMasterVolume(out level);
+            Marshal.ReleaseComObject(volume);
+            return level * 100;
+        }
+
+        public static bool? GetApplicationMute(int pid)
+        {
+            ISimpleAudioVolume? volume = GetVolumeObject(pid);
+            if (volume == null)
+                return null;
+
+            bool mute;
+            volume.GetMute(out mute);
+            Marshal.ReleaseComObject(volume);
+            return mute;
+        }
+
+        public static void SetApplicationVolume(int pid, float level)
+        {
+            ISimpleAudioVolume? volume = GetVolumeObject(pid);
+            if (volume == null)
+                return;
+
+            Guid guid = Guid.Empty;
+            volume.SetMasterVolume(level / 100, ref guid);
+            Marshal.ReleaseComObject(volume);
+        }
+
+        public static void SetApplicationMute(int pid, bool mute)
+        {
+            ISimpleAudioVolume? volume = GetVolumeObject(pid);
+            if (volume == null)
+                return;
+
+            Guid guid = Guid.Empty;
+            volume.SetMute(mute, ref guid);
+            Marshal.ReleaseComObject(volume);
+        }
+
+        private static ISimpleAudioVolume? GetVolumeObject(int pid)
+        {
+            IMMDeviceEnumerator? deviceEnumerator = null;
+            IAudioSessionEnumerator? sessionEnumerator = null;
+            IAudioSessionManager2? mgr = null;
+            IMMDevice? speakers = null;
+            try
+            {
+                // get the speakers (1st render + multimedia) device
+                deviceEnumerator = (IMMDeviceEnumerator)(new MMDeviceEnumerator());
+                deviceEnumerator.GetDefaultAudioEndpoint(EDataFlow.eRender, ERole.eMultimedia, out speakers);
+
+                // activate the session manager. we need the enumerator
+                Guid IID_IAudioSessionManager2 = typeof(IAudioSessionManager2).GUID;
+                object o = null;
+                speakers?.Activate(ref IID_IAudioSessionManager2, 0, IntPtr.Zero, out o);
+                mgr = o == null ? null : (IAudioSessionManager2)o;
+
+                // enumerate sessions for on this device
+                if (mgr == null)
+                {
+                    return null;
+                }
+                mgr.GetSessionEnumerator(out sessionEnumerator);
+                int count;
+                sessionEnumerator.GetCount(out count);
+
+                // search for an audio session with the required process-id
+                ISimpleAudioVolume? volumeControl = null;
+                for (int i = 0; i < count; ++i)
+                {
+                    IAudioSessionControl2? ctl = null;
+                    if (ctl == null)
+                    {
+                        continue;
+                    }
+
+                    try
+                    {
+                        sessionEnumerator.GetSession(i, out ctl);
+
+                        // NOTE: we could also use the app name from ctl.GetDisplayName()
+                        int cpid;
+                        
+                        ctl.GetProcessId(out cpid);
+
+                        if (cpid == pid)
+                        {
+                            volumeControl = (ISimpleAudioVolume)ctl;
+                            break;
+                        }
+                    }
+                    finally
+                    {
+                        if (ctl != null) Marshal.ReleaseComObject(ctl);
+                    }
+                }
+
+                return volumeControl;
+            }
+            finally
+            {
+                if (sessionEnumerator != null) Marshal.ReleaseComObject(sessionEnumerator);
+                if (mgr != null) Marshal.ReleaseComObject(mgr);
+                if (speakers != null) Marshal.ReleaseComObject(speakers);
+                if (deviceEnumerator != null) Marshal.ReleaseComObject(deviceEnumerator);
+            }
+        }
+
+        #endregion
+    }
+}
