+++ ExternalHelpers/WindowsMasterVolume/EDataFlow.cs
@@ -1,0 +1,10 @@
+﻿namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    internal enum EDataFlow
+    {
+        eRender,
+        eCapture,
+        eAll,
+        EDataFlow_enum_count
+    }
+}
