+++ ExternalHelpers/WindowsMasterVolume/ERole.cs
@@ -1,0 +1,10 @@
+﻿namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    internal enum ERole
+    {
+        eConsole,
+        eMultimedia,
+        eCommunications,
+        ERole_enum_count
+    }
+}
