+++ ExternalHelpers/WindowsMasterVolume/IAudioSessionEnumerator.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.InteropServices;
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    [Guid("E2F5BB11-0570-40CA-ACDD-3AA01277DEE8"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    internal interface IAudioSessionEnumerator
+    {
+        [PreserveSig]
+        int GetCount(out int SessionCount);
+
+        [PreserveSig]
+        int GetSession(int SessionCount, out IAudioSessionControl2? Session);
+    }
+}
