+++ ExternalHelpers/WindowsMasterVolume/IAudioSessionManager2.cs
@@ -1,0 +1,15 @@
+﻿using System.Runtime.InteropServices;
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    [Guid("77AA99A0-1BD6-484F-8BC7-2C654C9A9B6F"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    internal interface IAudioSessionManager2
+    {
+        int NotImpl1();
+        int NotImpl2();
+
+        [PreserveSig]
+        int GetSessionEnumerator(out IAudioSessionEnumerator SessionEnum);
+
+        // the rest is not implemented
+    }
+}
