+++ ExternalHelpers/WindowsMasterVolume/IMMDevice.cs
@@ -1,0 +1,12 @@
+﻿using System.Runtime.InteropServices;
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    [Guid("D666063F-1587-4E43-81F1-B948E807363F"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    internal interface IMMDevice
+    {
+        [PreserveSig]
+        int Activate(ref Guid iid, int dwClsCtx, IntPtr pActivationParams, [MarshalAs(UnmanagedType.IUnknown)] out object? ppInterface);
+
+        // the rest is not implemented
+    }
+}
