+++ ExternalHelpers/WindowsMasterVolume/ISimpleAudioVolume.cs
@@ -1,0 +1,19 @@
+﻿using System.Runtime.InteropServices;
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    [Guid("87CE5498-68D6-44E5-9215-6DA47EF883D8"), InterfaceType(ComInterfaceType.InterfaceIsIUnknown)]
+    internal interface ISimpleAudioVolume
+    {
+        [PreserveSig]
+        int SetMasterVolume(float fLevel, ref Guid EventContext);
+
+        [PreserveSig]
+        int GetMasterVolume(out float pfLevel);
+
+        [PreserveSig]
+        int SetMute(bool bMute, ref Guid EventContext);
+
+        [PreserveSig]
+        int GetMute(out bool pbMute);
+    }
+}
