+++ ExternalHelpers/WindowsMasterVolume/MMDeviceEnumerator.cs
@@ -1,0 +1,9 @@
+﻿using System.Runtime.InteropServices;
+namespace PowerControl.Helpers.WindowsMasterVolume
+{
+    [ComImport]
+    [Guid("BCDE0395-E52F-467C-8E3D-C4579291692E")]
+    internal class MMDeviceEnumerator
+    {
+    }
+}
